analytics_service/requirements.txt
amqp==5.3.1
anyio==4.12.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
django==6.0.2
djangorestframework==3.16.1
elastic-transport==9.2.1
elasticsearch==9.3.0
gunicorn==25.0.1
idna==3.11
kombu==5.6.2
packaging==26.0
pika==1.3.2
prompt-toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==7.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
typing-extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.5.3
